MAGNeto/requirements.txt
numpy==1.19.0
scipy==1.5.2
opencv-python==4.2.0.34
pandas==1.0.5
Pillow==6.1.0
tensorboard==2.2.2
torch==1.5.1
torchvision==0.6.1
tqdm==4.47.0
matplotlib==3.2.2
